requirements.txt
altgraph==0.17.3
black==22.12.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
docx==0.2.4
idna==3.4
lxml==4.9.1
macholib==1.16.2
mypy-extensions==0.4.3
pathspec==0.10.3
Pillow==9.2.0
platformdirs==2.6.2
pyinstaller==5.8.0
pyinstaller-hooks-contrib==2023.0
PySimpleGUI==4.60.4
python-pptx==0.6.19
six==1.16.0
tomli==2.0.1
urllib3==1.26.13
XlsxWriter==3.0.3
